pyproject.toml
[project]
name = "fastapi-rbac-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb==0.4.15",
    "fastapi==0.103.1",
    "groq==0.4.1",
    "huggingface-hub==0.15.1",
    "langchain==0.0.306",
    "langchain-groq==0.0.1",
    "oso==0.27.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]==2.3.0",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "sentence-transformers==2.2.2",
    "sqlalchemy==2.0.20",
    "tqdm==4.66.1",
    "uvicorn==0.23.2",
]
